requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
python-decouple==3.8
Pillow==10.0.1
django-extensions==3.2.3
djangorestframework-nested==0.0.1
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1
openai==1.3.5
python-magic==0.4.27
reportlab==4.0.4
requests==2.31.0 